+++ client/src/pages/import-short-tests/components/adjust-generated-short-test/AdjustGeneratedShortTest.tsx
@@ -4,9 +4,8 @@ import _ from 'lodash';
 import React, { useCallback, useMemo, useState } from 'react';
 import { IExerciseGradingDTO, IGradingDTO } from 'shared/model/Gradings';
 import ShortTestForm, { ShortTestFormState } from '../../../../components/forms/ShortTestForm';
-import { useImportCSVContext } from '../../../../components/import-csv-new/ImportCSV.context';
-import { CSVMappedColumns } from '../../../../components/import-csv-new/ImportCSV.types';
-import { CSVDataRow } from '../../../../components/import-csv/ImportCSV.types';
+import { useImportCSVContext } from '../../../../components/import-csv/ImportCSV.context';
+import { CSVDataRow, CSVMappedColumns } from '../../../../components/import-csv/ImportCSV.types';
 import LoadingModal from '../../../../components/loading/LoadingModal';
 import Placeholder from '../../../../components/Placeholder';
 import HookUpStepperWithFormik from '../../../../components/stepper-with-buttons/HookUpStepperWithFormik';
@@ -19,7 +18,7 @@ import { Student } from '../../../../model/Student';
 import { FormikSubmitCallback } from '../../../../types';
 import { Logger } from '../../../../util/Logger';
 import { convertFormStateToDTO } from '../../../hand-ins/short-test-management/ShortTestManagement';
-import { ShortTestColumns } from '../../ImportShortTests';
+import { ShortTestColumns } from '../ImportShortTests';
 import { useIliasMappingContext } from '../map-students-ilias-names/IliasMapping.context';
 import { generateInitialValues, validateExercises, validateShortTestNumber } from './formHelpers';
